rg_i2c: rg_i2c_read_byte can now indicate failure by returning -1

+++ components/retro-go/rg_i2c.h
@@ -8,7 +8,7 @@ bool rg_i2c_init(void);
 bool rg_i2c_deinit(void);
 bool rg_i2c_read(uint8_t addr, int reg, void *read_data, size_t read_len);
 bool rg_i2c_write(uint8_t addr, int reg, const void *write_data, size_t write_len);
-uint8_t rg_i2c_read_byte(uint8_t addr, uint8_t reg);
+int rg_i2c_read_byte(uint8_t addr, uint8_t reg);
 bool rg_i2c_write_byte(uint8_t addr, uint8_t reg, uint8_t value);
